Finishing touches.
fix details on css to make it look like original. bit of a hack job in
many spots but i'm a bit burned out on this project which I dont even
like so I think it's time to move on to something new and exciting to
polish my skills. Place special atention to setting margins properly
so I can take advantage of flex properties.